Implement oauth authentication with google as the provider
Added a login page with a google sign in button, which displays
a login form when you click on it. If the user allows the app
to connect, the callback function signInCallback is called, which
sends an authorization code provided by google back to '/gconnect'.

The server validates the authorization code with the google API,
and collects some information about the user such as their email,
profile picture link, and user name.

The gdisconnect page uses the google oauth API to revoke the access
token provided, thereby logging out the user.